requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.1
click==8.1.7
decorator==5.1.1
Django==4.2.25
django-environ==0.11.2
djangorestframework==3.15.2
drf-spectacular==0.27.2
et-xmlfile==1.1.0
Flask==3.0.3
idna==3.10
influxdb-client==1.46.0
itsdangerous==2.2.0
Jinja2==3.1.4
kafka-python==2.0.2
MarkupSafe==2.1.5
modbus-tk==1.1.3
numpy==1.24.4
openpyxl==3.1.5
paho-mqtt==2.1.0
pandas==2.0.3
py==1.11.0
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2024.2
reactivex==4.0.4
redis==5.0.7
requests==2.32.3
retry==0.9.2
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
celery==5.4.0
